+++ src/main/java/br/unitins/tp1/monitores/service/user/UsuarioServiceImpl.java
@@ -18,25 +18,45 @@ public class UsuarioServiceImpl implements UsuarioService {
 
     @Inject
     public UsuarioRepository usuarioRepository;
+    
+
+    
+
 
     @Override
     public Usuario findById(Long id) {
+        if(id == null) {
+            throw new ValidationException( "id", "Id não pode ser nulo");
+        }
         return usuarioRepository.findById(id);
     }
 
     @Override
     public Usuario findByUsernameAndSenha(String username, String senha) {
+        if(username == null) {
+            throw new ValidationException( "username", "username não pode ser nulo");
+        }
+        if(senha == null) {
+            throw new ValidationException( "senha", "senha não pode ser nulo");
+        }
         return usuarioRepository.findByUsernameAndSenha(username, senha);
     }
 
     @Override
     public List<Usuario> findByNome(String nome) {
+        if(nome == null) {
+            throw new ValidationException( "nome", "nome não pode ser nulo");
+        }
         return null;
     }
 
     @Override
     public List<Usuario> findAll() {
-        return usuarioRepository.findAll().list();
+        try {
+            return usuarioRepository.findAll().list();
+        } catch (Exception e) {
+            throw new ValidationException("findAll", "Erro ao buscar usuários: " + e.getMessage());
+        }
     }
 
     @Inject
@@ -45,10 +65,24 @@ public class UsuarioServiceImpl implements UsuarioService {
     @Override
     @Transactional
     public Usuario create(UsuarioCreateRequestDTO dto) {
+
+        if (dto.username() == null || dto.username().isBlank()) {
+            throw new ValidationException("username", "Username é obrigatório.");
+        }
+    
+        if (usuarioRepository.findByUsername(dto.username()) != null) {
+            throw new ValidationException("username", "Username já cadastrado.");
+        }
+    
+    
+        if (dto.senha() == null || dto.senha().isBlank()) {
+            throw new ValidationException("senha", "Senha é obrigatória.");
+        }
+
         Usuario usuario = new Usuario();
         usuario.setUsername(dto.username());
-        usuario.setSenha(hashService.getHashSenha(dto.senha())); // Hash da senha
-        usuario.setPerfil(Perfil.valueOf(dto.perfil())); // Converta o perfil para o enum
+        usuario.setSenha(hashService.getHashSenha(dto.senha()));
+        usuario.setPerfil(Perfil.USER); 
 
         usuarioRepository.persist(usuario);
         return usuario;
@@ -62,32 +96,35 @@ public class UsuarioServiceImpl implements UsuarioService {
             throw new ValidationException("usuario", "Usuário não encontrado");
         }
 
-        // Verifica se a senha antiga está correta
         String hashSenhaAntiga = hashService.getHashSenha(dto.senhaAntiga());
         if (!usuario.getSenha().equals(hashSenhaAntiga)) {
             throw new ValidationException("senhaAntiga", "A senha antiga está incorreta");
         }
 
-        // Verifica se o username antigo está correto
         if (!usuario.getUsername().equals(dto.usernameAntigo())) {
             throw new ValidationException("usernameAntigo", "O username antigo está incorreto");
         }
 
-        // Atualiza o username e a senha
-        usuario.setUsername(dto.novoUsername()); // Atualiza o username
-        usuario.setSenha(hashService.getHashSenha(dto.novaSenha())); // Atualiza a senha com hash
+        usuario.setUsername(dto.novoUsername()); 
+        usuario.setSenha(hashService.getHashSenha(dto.novaSenha())); 
 
-        return usuario; // O usuário é atualizado automaticamente pelo Panache
+        return usuario;
     }
 
     @Override
     @Transactional
     public void delete(Long id) {
+        if(id == null) {
+            throw new ValidationException( "id", "Id não pode ser nulo");
+        }
         usuarioRepository.deleteById(id);
     }
 
     @Override
     public Usuario findByUsername(String username) {
+        if (username == null) {
+            throw new ValidationException("username", "Username não pode ser nulo");
+        }
         return usuarioRepository.findByUsername(username);
     }
